feat: 기록하기 뷰에서 수정하기 뷰 관련 내용 제거
