+++ app/src/main/java/com/team/recordream/presentation/MainActivity.kt
@@ -45,15 +45,7 @@ class MainActivity : BindingActivity<ActivityMainBinding>(R.layout.activity_main
 
     private fun setClickEvents() {
         binding.ivMainMypage.setOnClickListener { startActivity(MypageActivity.getIntent(this)) }
-        binding.ivNaviWriteBtn.setOnClickListener {
-            startActivity(
-                RecordActivity.getIntent(
-                    this,
-                    RecordActivity.CREATE_MODE,
-                    null,
-                ),
-            )
-        }
+        binding.ivNaviWriteBtn.setOnClickListener { startActivity(RecordActivity.getIntent(this)) }
         binding.ivMainSearch.setOnClickListener { startActivity(SearchActivity.getIntent(this)) }
     }
