+++ app/src/main/res/layout/activity_record.xml
@@ -35,6 +35,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="32dp"
+            android:text="@string/tv_record_create"
             android:textAppearance="@style/Sb16"
             android:textColor="@color/white01_FFFFFF"
             app:layout_constraintEnd_toEndOf="parent"
